Fixed MSVC compilation error
MSVC cannot recognize some special characters in RapidJSON error message C string array in consts.hpp. This is really hard to find.

+++ inc/floatint.hpp
@@ -26,7 +26,7 @@ namespace FloatInt
   {
     private:
       int data;
-      unsigned int digits;
+      size_t digits;
     public:
       int getData() const { return data; }
       int getDigits() const { return digits; }
